Allow update callback to have $this
The *update* callback should have a $this context, or argument passed so it is easier to use it to traverse, and find another element to which _current_percentage_ can be displayed. This could be used when multiple progressbars are used on the same page, each placed in it's logical group.

The change is very trivial, plus to avoid breaking any existing users, I chose to add $this as a 2nd argument, although personally I'd prefer to pass $this as the first argument.