Additions to bower.json
This package is a JQuery plugin, and I believe it depends on JQuery to be installed. I think it would be correct to add 'dependencies' settings into bower.json indicating this fact. If it also depends on Bootstrap (its JS part that is) please feel free to add that as a dependency as well.